fix(controllers): Fix model name for E2213 in ZHA and deconz
This PR fixes the capitalization of the `model` property of the E2213
blueprint for the ZHA and deconz integrations. I have tested the
original form with ZHA and got "no matching models found." With only
lower case `shortcut button`, the model is found. Note that that matches
the `model` used for MQTT. I only *assume* that the model name is the
lower-case variant in deconz as well; it may be worth checking, leaving
as is, or have both variants instead.